fix(replication): full-snapshot despawn no longer truncates a 64-bit NetId
ClientReplicationView's full-state despawn loop iterated the gone List<long>
as `int`, so an id past 2^31 leaving the snapshot truncated: it threw
KeyNotFoundException out of the apply (killing the session) or, on a low-bits
collision, despawned the wrong entity while the departed one ghosted. Latent
since 10.0.0 (single-node ids are small). Fixed to `long`; the delta path was
already correct. Added despawn-path boundary tests (0 / int.MaxValue / >2^32 /
high-node-bits) plus the collision case (5 vs Pack(1,5)); the shipped
NetIdWireBoundaryTests never applied a second snapshot with the entity absent.

+++ KhaozEngine.Replication/ClientReplicationView.cs
@@ -76,7 +76,7 @@ public sealed class ClientReplicationView
             if (!seen.Contains(kv.Key)) (gone ??= new List<long>()).Add(kv.Key);
         if (gone is not null)
         {
-            foreach (int netId in gone)
+            foreach (long netId in gone)   // long, not int: a full-snapshot despawn must not truncate a 64-bit id
             {
                 if (world.IsAlive(entityByNetId[netId])) world.Despawn(entityByNetId[netId]);
                 entityByNetId.Remove(netId);
